Migrate beacon to dom
